+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # ChangeLog
 
 # V1.13.0-pre9
+* Backed out change where we try "monitor exit". See [Issue #1185](https://github.com/Marus/cortex-debug/issues/1185)
+* Increased server temout to 10 mins
 * Fix for https://github.com/Marus/cortex-debug/issues/1167 Race comdition in parallel reading of multiple symbol files
 
 # V1.13.0-pre8